+++ bluetooth_low_energy_android/lib/src/my_bluetooth_low_energy.dart
@@ -12,10 +12,4 @@ class MyBluetoothLowEnergy extends BluetoothLowEnergy {
   MyBluetoothLowEnergy()
       : centralManager = MyCentralManager(),
         peripheralManager = MyPeripheralManager();
-
-  @override
-  Future<void> setUp() async {
-    await centralManager.setUp();
-    await peripheralManager.setUp();
-  }
 }