+++ bluetooth_low_energy_linux/lib/src/my_bluetooth_low_energy.dart
@@ -9,9 +9,4 @@ class MyBluetoothLowEnergy extends BluetoothLowEnergy {
   PeripheralManager get peripheralManager => throw UnimplementedError();
 
   MyBluetoothLowEnergy() : centralManager = MyCentralManager();
-
-  @override
-  Future<void> setUp() async {
-    await centralManager.setUp();
-  }
 }