+++ bluetooth_low_energy_platform_interface/lib/src/bluetooth_low_energy_manager.dart
@@ -8,4 +8,7 @@ abstract class BluetoothLowEnergyManager {
 
   /// Tells the manager’s state updated.
   Stream<BluetoothLowEnergyStateChangedEventArgs> get stateChanged;
+
+  /// Sets up this bluetooth low energy manager.
+  Future<void> setUp();
 }